requirements.txt
pytest==6.2.2  
selenium==3.141.0 
webdriver_manager==3.3.0 
allure-pytest==2.9.43
allure-python-commons==2.9.43
atomicwrites==1.4.0
attrs==21.2.0
b==1.0.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
et-xmlfile==1.1.0
idna==2.10
iniconfig==1.1.1
logging-logger==0.1
numpy==1.21.0
openpyxl==3.0.7
packaging==21.0
pandas==1.3.0
pip==21.1.3
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
selenium==3.141.0
setuptools==57.1.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.6
webdriver-manager==3.4.2
pytest-xdist